refactor: Clean up KnoraConstants
