+++ projects/knora/core/src/lib/declarations/api/admin/lists/list-node-update-payload.ts
@@ -1,4 +1,4 @@
-import { StringLiteral } from '../../../';
+import { StringLiteral } from '@knora/api';
 
 /**
  * @deprecated Use new model from `@knora/api` (github:dasch-swiss/knora-api-js-lib) instead
