+++ projects/knora/core/src/lib/declarations/api/admin/projects/project.ts
@@ -1,5 +1,5 @@
 import { JsonObject, JsonProperty } from 'json2typescript';
-import { StringLiteral } from '../../shared/strings';
+import { StringLiteral } from '@knora/api';
 
 /**
  * @deprecated Use new model from `@knora/api` (github:dasch-swiss/knora-api-js-lib) instead
