+++ projects/knora/core/src/public_api.ts
@@ -6,6 +6,6 @@ export * from './lib/core.module';
 export * from './lib/core.config';
 export * from './lib/session.service';
 
-export * from './lib/declarations/';
-export * from './lib/declarations/api/operators';
+// export * from './lib/declarations/';
+// export * from './lib/declarations/api/operators';
 export * from './lib/services/';
